requirements.txt
jax
jaxlib
absl-py
array-record
flax@git+https://github.com/wsmoses/flax@d854bf3c29c5048c1f2265d3659ba995bf218476
ml-collections
numpy
optax
protobuf # ==3.20.3
pylint
pytest
pytype
sentencepiece==0.1.97
tensorflow-text; python_version < '3.12'
tensorflow
tensorflow-datasets
tiktoken
transformers
google-jetstream
